feat: Add comprehensive tests for PaymentController
Implements US-TEST-1.

Adds a new feature test suite for the PaymentController to ensure the critical payment process is robust and reliable.

The suite covers all scenarios for the 'create' and 'store' methods, including authorization, offer status checks, successful payments (card, wallet, mixed), and various failure cases.

A minor bug fix was applied to the controller to handle type comparison correctly during amount verification.

Existing tests that were affected by this change have been updated to prevent regressions.

+++ pifpaf/app/Http/Controllers/PaymentController.php
@@ -99,7 +99,7 @@ class PaymentController extends Controller
             }
 
             // Vérification de sécurité : le montant payé correspond-il au montant attendu ?
-            $expectedAmountInCents = round($cardAmount * 100);
+            $expectedAmountInCents = (int) round($cardAmount * 100);
             if ($intent->amount !== $expectedAmountInCents) {
                 // Potentielle fraude ou erreur, on ne finalise pas
                 // On pourrait aussi rembourser le paiement ici
